requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
google-genai==1.44.0
python-dotenv==1.0.0
websockets==15.0.1
requests==2.32.5
google-cloud-recaptcha-enterprise==1.19.0
google-cloud-bigquery==3.30.0
redis==7.1.0
slowapi==0.1.9
starlette==0.45.3
PyJWT==2.8.0
